- summing first green rake 0 and 90 green functions before fft_reconvolve_stf reduces call to this function (faster).
- loading all the hdf5 content to a dictionnary makes also things much faster.

processing 10 files, 100 sources each, from:

real    0m52,284s
user    0m50,512s
sys     0m11,025s

to:

real    0m30,747s
user    0m29,612s
sys     0m10,437s